requirements.txt
pydantic
python-dotenv
langchain_groq
langchain-core
langgraph
streamlit
markdown
